Build Updates

+++ build.fsx
@@ -18,9 +18,9 @@ let isPR = match isAppVeyor with
            | true-> not(String.IsNullOrWhiteSpace AppVeyorEnvironment.PullRequestNumber)
            | _ -> false
                  
-let isStable = not(isPR) && branchName = "master"
-let isBeta = not(isPR) && (isStable || branchName = "release")
-let isAlpha = not(isPR) && (isBeta || branchName = "develop")
+let isStable = branchName = "master"
+let isBeta = (isStable || branchName = "release")
+let isAlpha = (isBeta || branchName = "develop")
 
 printfn "Running On Branch: %s" branchName
 printfn "PR Number: %s" AppVeyorEnvironment.PullRequestNumber
@@ -42,14 +42,18 @@ let versionNumber = match isStable with
                     | _ -> sprintf "%s.%s" baseVersion buildNumber
 
 Target "Clean" (fun _ -> 
+    printfn "Clean & Ensure Output Directory"
     CleanDir outputDir
 )
 
 Target "VersionUpdate" (fun _ ->
+    printfn "Update Assembly Info Versions"
     BulkReplaceAssemblyInfoVersions "src/" (fun f -> { f with AssemblyVersion = versionNumber; AssemblyInformationalVersion = versionNumber; AssemblyFileVersion = versionNumber })
-        
+    
+    printfn "Update Change Log Versions"
     File.WriteAllText(changeLogPath, File.ReadAllText(changeLogPath).Replace("<Version Number=\"0.0.0.0\" Name=\"Pre-Release\">", (sprintf "<Version Number=\"%s\" Name=\"Pre-Release\">" versionNumber)))
 
+    printfn "Update Click-Once Settings Versions"
     Directory.GetFiles(srcDir, "*proj", SearchOption.AllDirectories)
     |> Seq.iter(fun x -> File.WriteAllText(x, File.ReadAllText(x).Replace("<MinimumRequiredVersion>0.0.0.0</MinimumRequiredVersion>", (sprintf "<MinimumRequiredVersion>%s</MinimumRequiredVersion>" versionNumber))
                                                                  .Replace("<ApplicationVersion>0.0.0.0</ApplicationVersion>", (sprintf "<ApplicationVersion>%s</ApplicationVersion>" versionNumber))))
@@ -88,49 +92,53 @@ Target "Package" (fun _ ->
         File.Copy(outputDir @@ "Stable" @@ "setup.exe", outputDir @@ "stable-setup.exe")
 )
 
-Target "Publish" (fun _ ->
+Target "Publish-Artifacts" (fun _ ->
     PushArtifacts (Directory.GetFiles(outputDir, "*.zip", SearchOption.TopDirectoryOnly))
     PushArtifacts (Directory.GetFiles(outputDir, "*.exe", SearchOption.TopDirectoryOnly))
-    
-    let releasesRepo = outputDir @@ "Releases"
-    DeleteDir releasesRepo |> ignore
+)
 
+Target "Publish-Release" (fun _ ->    
+    let releasesRepo = outputDir @@ "Releases"
+
+    //Hide process tracing so the access token doesn't show
+    printfn "Downloading Releases Repo from GitHub"
     enableProcessTracing <- false
     let authKey = environVar "access_token"
     cloneSingleBranch outputDir (sprintf "https://%s:x-oauth-basic@github.com/BlythMeister/Gallifrey.Releases.git" authKey) "master" "Releases"
+    enableProcessTracing <- true
+
     directRunGitCommandAndFail releasesRepo "config --global user.email \"publish@gallifreyapp.co.uk\""
     directRunGitCommandAndFail releasesRepo "config --global user.name \"Gallifrey Auto Publish\""
 
-    let publishRelease (releaseType:string) = 
-        let sourceRoot = outputDir @@ releaseType
+    let publishRelease (releaseType:string) =         
         let destinationRoot = releasesRepo @@ "download" @@ "modern" @@ (releaseType.ToLower())
+        let destinationAppFilesRoot = destinationRoot @@ "Application Files"
+
         ensureDirectory destinationRoot
-        File.Copy(sourceRoot @@ (sprintf "Gallifrey.UI.Modern.%s.application" releaseType), destinationRoot @@ (sprintf "Gallifrey.UI.Modern.%s.application" releaseType), true)
+        match isStable, releaseType with
+        | true, "Alpha" 
+        | true, "Beta" -> CleanDir destinationAppFilesRoot
+        | _, _ -> ensureDirectory destinationAppFilesRoot 
+       
+        let destinationAppFiles = destinationAppFilesRoot @@ (sprintf "Gallifrey.UI.Modern.Alpha_%s" (versionNumber.Replace(".","_")))
 
-        let destinationFiles = destinationRoot @@ "Application Files"
-        ensureDirectory destinationFiles
-        let preCount = Directory.GetDirectories(destinationFiles) |> Seq.length
-        
-        Directory.GetDirectories(sourceRoot @@ "Application Files")
-        |> Seq.map(fun x -> new DirectoryInfo(x))
-        |> Seq.iter(fun x -> Directory.Move(x.FullName, destinationFiles @@ x.Name))
-        
-        let postCount = Directory.GetDirectories(destinationFiles) |> Seq.length
+        if not(Directory.Exists destinationAppFiles) then
+            printfn "Publish Relase Type: %s" releaseType
+            let sourceRoot = outputDir @@ releaseType
 
-        if postCount > preCount then
+            let sourceAppFiles = sourceRoot @@ "Application Files" @@ (sprintf "Gallifrey.UI.Modern.Alpha_%s" (versionNumber.Replace(".","_")))
+            File.Copy(sourceRoot @@ (sprintf "Gallifrey.UI.Modern.%s.application" releaseType), destinationRoot @@ (sprintf "Gallifrey.UI.Modern.%s.application" releaseType), true)
+            Directory.Move(sourceAppFiles, destinationAppFiles)
+            
             StageAll releasesRepo
             Commit releasesRepo (sprintf "Publish %s - %s" releaseType versionNumber)
-            
-            if isStable then
-                pushTag currentDirectory "origin" baseVersion
-
+            if isStable then pushTag currentDirectory "origin" baseVersion
+                           
     if isAlpha then publishRelease "Alpha"
     if isBeta then publishRelease "Beta"
     if isStable then publishRelease "Stable"
 
-    pushBranch releasesRepo "origin" "master"
-    
-    enableProcessTracing <- true
+    pushBranch releasesRepo "origin" "master"   
 )
 
 Target "Default" DoNothing
@@ -139,7 +147,8 @@ Target "Default" DoNothing
     ==> "VersionUpdate"
     ==> "Build"
     ==> "Package"
-    =?> ("Publish", isAppVeyor && (isAlpha || isBeta || isStable))
+    =?> ("Publish-Artifacts", isAppVeyor)
+    =?> ("Publish-Release", isAppVeyor && not(isPR) && (isAlpha || isBeta || isStable))
     ==> "Default"
 
 RunParameterTargetOrDefault "target" "Default"
